requirements.txt
streamlit
pandas
numpy
matplotlib
google-cloud-bigquery==2.34.4
google-auth==2.23.4